Small updates Save as Tiff, GLDAS, GLEAM, MCD43
Save as tiff now also creates a directory.
GLDAS an empty line is removed in the data.
SoilGrids the layer of the carbon content is added in filename

+++ Collect/MCD43/DataAccess.py
@@ -117,29 +117,33 @@ def RetrieveData(Date, args):
     # Argument
     [output_folder, TilesVertical, TilesHorizontal, lonlim, latlim, hdf_library] = args
 
-    # Collect the data from the MODIS webpage and returns the data and lat and long in meters of those tiles
-    try:
-        Collect_data(TilesHorizontal, TilesVertical, Date, output_folder, hdf_library)
-    except:
-        print("Was not able to download the file")
-
-      # Define the output name of the collect data function
-    name_collect = os.path.join(output_folder, 'Merged.tif')
-
-    # Reproject the MODIS product to epsg_to
-    epsg_to ='4326'
-    name_reprojected = RC.reproject_MODIS(name_collect, epsg_to)
-
-    # Clip the data to the users extend
-    data, geo = RC.clip_data(name_reprojected, latlim, lonlim)
-
-    # Save results as Gtiff
+    # output filename
     ReffileName = os.path.join(output_folder, 'Albedo_MCD43A3_-_daily_' + Date.strftime('%Y') + '.' + Date.strftime('%m') + '.' + Date.strftime('%d') + '.tif')
-    DC.Save_as_tiff(name=ReffileName, data=data, geo=geo, projection='WGS84')
 
-    # remove the side products
-    os.remove(os.path.join(output_folder, name_collect))
-    os.remove(os.path.join(output_folder, name_reprojected))
+    if not os.path.exists(ReffileName):
+        
+        # Collect the data from the MODIS webpage and returns the data and lat and long in meters of those tiles
+        try:
+            Collect_data(TilesHorizontal, TilesVertical, Date, output_folder, hdf_library)
+        except:
+            print("Was not able to download the file")
+    
+          # Define the output name of the collect data function
+        name_collect = os.path.join(output_folder, 'Merged.tif')
+    
+        # Reproject the MODIS product to epsg_to
+        epsg_to ='4326'
+        name_reprojected = RC.reproject_MODIS(name_collect, epsg_to)
+    
+        # Clip the data to the users extend
+        data, geo = RC.clip_data(name_reprojected, latlim, lonlim)
+    
+        # Save results as Gtiff
+        DC.Save_as_tiff(name=ReffileName, data=data, geo=geo, projection='WGS84')
+    
+        # remove the side products
+        os.remove(os.path.join(output_folder, name_collect))
+        os.remove(os.path.join(output_folder, name_reprojected))
 
     return True
